// apps/hms-api/src/@generated/prisma-nestjs-graphql/prisma/float-filter.input.ts
import { Field, InputType, Float } from '@nestjs/graphql';

@InputType()
export class FloatFilter {
  @Field(() => Float, { nullable: true })
  equals?: number;

  @Field(() => [Float], { nullable: true })
  in?: Array<number>;

  @Field(() => [Float], { nullable: true })
  notIn?: Array<number>;

  @Field(() => Float, { nullable: true })
  lt?: number;

  @Field(() => Float, { nullable: true })
  lte?: number;

  @Field(() => Float, { nullable: true })
  gt?: number;

  @Field(() => Float, { nullable: true })
  gte?: number;

  @Field(() => FloatFilter, { nullable: true })
  not?: FloatFilter;
}